requirements.txt
torch==1.12.1
numpy==1.21.2
matplotlib==3.5.0
scikit-image==0.18.3
pillow==8.4.0
timm==0.4.12
transformers==4.24.0
torchtext==0.13.1
torchvision>=0.11.1
datasets==2.6.1
